Middleware-reconfiguration: Implemented classes and renamed files
